Move JiraUtil to separate package
