+++ src/main/java/org/embulk/input/jira/JiraInputPlugin.java
@@ -6,6 +6,7 @@ import org.embulk.config.ConfigSource;
 import org.embulk.config.Task;
 import org.embulk.config.TaskReport;
 import org.embulk.config.TaskSource;
+import org.embulk.input.jira.util.JiraUtil;
 import org.embulk.spi.Exec;
 import org.embulk.spi.InputPlugin;
 import org.embulk.spi.PageOutput;
